jwt 6
Added authorization headers and token to signin component.  The function adds the token to the session storage if the user signs in successfully.  Added authorization check to app component.  The auth function checks if the token exists in session storage and if so saves the user data to state and routes the user to the home page.