jwt 8
added age and pet fields to the user schema in database.  Updated profile and app components to include the new fields when submittied in the form and display them on the front end.  Added signout functionality to app component.  The function nullifies the token in session storage and sends a null authorization to the signin route.